// web-src/math/fork/src/public/core-types.ts
export type DimensionUnit =
  | 'pt'
  | 'mm'
  | 'cm'
  | 'ex'
  | 'px'
  | 'em'
  | 'bp'
  | 'dd'
  | 'pc'
  | 'in'
  | 'mu'
  | 'fil'
  | 'fill'
  | 'filll';

/**
 * A dimension is used to specify the size of things
 *
 * @category Registers
 */
export type Dimension = {
  dimension: number;
  unit?: DimensionUnit; // If missing, assumes 'pt'
};
// ...
export type MatrixEnvironment =
  | 'matrix'
  | 'matrix*'
  | 'pmatrix'
  | 'pmatrix*'
  | 'bmatrix'
  | 'bmatrix*'
  | 'Bmatrix'
  | 'Bmatrix*'
  | 'vmatrix'
  | 'vmatrix*'
  | 'Vmatrix'
  | 'Vmatrix*';

/** @internal */
export type CasesEnvironment = 'cases' | 'dcases' | 'rcases';

/** @internal */
export type TabularEnvironment =
  | 'array'
  | 'equation'
  | 'equation*'
  | 'subequations'
  | 'multline'
  | 'align'
  | 'align*'
  | 'alignat'
  | 'alignat*'
  | 'flalign'
  | 'flalign*'
  | 'aligned'
  | 'eqnarray'
  | 'split'
  | 'gather'
  | 'gather*'
  | 'gathered'
  | 'lines'
  | 'multline'
  | 'multline*'
  | 'cases'
  | 'dcases'
  | 'rcases'
  | 'smallmatrix'
  | 'smallmatrix*'
  | CasesEnvironment
  | MatrixEnvironment;

/** @internal */
export type AlignEnvironment =
  | 'align'
  | 'align*'
  | 'alignat'
  | 'alignat*'
  | 'flalign'
  | 'flalign*'
  | 'aligned'
  | 'gather'
  | 'gather*'
  | 'gathered'
  | 'split'
  | 'multline'
  | 'multline*';

/** @internal */
export type Environment =
  | 'math'
  | 'displaymath'
  | 'center'
  | TabularEnvironment;